1、增加升级的权限判断

+++ app/src/main/java/com/xinze/xinze/http/DownloadHandler.java
@@ -76,7 +76,7 @@ class DownloadHandler {
         Message message = mHandler.obtainMessage();
         message.what = WHAT_UPDATE;
         Bundle data = new Bundle();
-        data.putFloat(PROGRESS, progress);
+        data.putInt(PROGRESS, progress);
         message.setData(data);
         mHandler.sendMessage(message);
     }
